Change struct separator to align with JavaScript
To me, the benefit of using a colon (`:`) as a struct separator is it makes the JavaScript interop easier.  The default formatting of objects in popular JavaScript formatters is `{ key: value }`. (In fact, there is [no option to change this](https://prettier.io/docs/en/options.html) in Prettier.)  I propose we follow that standard.